requirements.txt
#pip install -r requirements.txt
django
pillow
django-crispy-forms
djangorestframework
cx_oracle
django-pwa